Fixes in 2018/poikola
I had put in the README.md that the entry needs XQuartz for macOS
because it uses X11 but this is not true (I should have known as this
was my favourite entry that year in the preview and it still is). The
point was that Terminal.app of macOS might not work. This has been
corrected.

As it happens it appears to work fine for me which I noted in the
README.md.

Formatting fixes, slight improvements and possibly typo fixes.

Add missing make rule docs that was referred to in the README.md but was
missing from the Makefile. It was found in the
archive/2020/poikola/Makefile.