Minor fixes in 2018/algmyr
Typo / formatting fixes.

I did not run the command to see if it works because I had a cat
sleeping between my legs. He just got up but I need to do other things
now, having finished for the first pass the year 2018.

The author noted a segfault on invalid files and I might look at
addressing this but this will come later if I do (likely but probably
will move the original source to prog.alt.c as I love the format so
much).