Fix 1988/isaak for modern systems
The problem was the exit() (redefined) was not being called in main().
This used to work but no longer does.

This fix was based on the file on the IOCCC website isaak.ansi.c. One of
the fixes in that one was the fact that in the older days one could
define define to something else and use that new macro as a define. But
since this was already done I didn't see the point in doing it again (it
was already in the repo as the diff shows - isaak.alt.c is the
original).

Removed this entry from bugs.md.