Explain magic of 1994/schnitzi in bugs.md
Since the entry cannot yet compile if it's changed to use fgets() I
decided to take a look at the source. Quite quickly I noticed something
fun at the top. If you look at each column and go down row by row you
see it spelling out the code! I wrote the following tips in the bugs.md
as an example:

    Take a look at the interesting comment as well as the `int
    r=0,x,y=0` at the top of the file. If you look at each column go
    down that column you can see how it spells out the code. For
    instance the first column looks like:

	i
	n
	c
	l
	u
	d
	e

	<
	s
	t
	d
	i
	o
	.
	h
	>

    If you join the lines you end up with:

	#include <stdio.h>

    If you look at column 25 which is the end of the word 'mh111' and
    you go down to the next row you'll see a 0 and if you go one row
    down another 0. This is the buffer size, 100, for `u`. The column to
    the left is the same for the `t` variable.

    Thus it seems that in order to get the generated output correct one
    needs to provide the correct input in comments or possibly by
    rearranging some of the code (this was required to make the
    generated code compile at all when changing the buffer size).

From this it should be possible, theoretically, to come up with the
right code so that everything works okay.