Fixes in 2019/burton/Makefile and README.md
The Makefile had a bad character, a '%' instead of a '$' which caused a
rule to fail.

Format fixes in README.md.

There is one thing I'm not sure about. It appears to be a mistake but
I'm not sure if there is some meaning. Line 66:

    Input should be less than ten million octets toavoidthisproblem.

But the last 'word' should be, at least if it is a mistake, four words.
I have not changed it as the context was not clear enough to be sure it
was incorrect. This can be dealt with another time.