Update faq on rule 2 wrt fixing entries
It's very possible that in fixing entries rule 2 will be broken in some
entries. This should be noted in the faq as people might notice it and
find it dubious. The new entry now points out that when fixing entries
it's very possible that rule 2 could inadvertently be broken but the
originals can be seen in /archive.