Fixes in 2019/endoh
Remove optimisation from Makefile and add -g to compile line as it's a
backtrace quine.

Add to .gitignore prog2.c as the author's comments suggest a command
that will create it.

I believe that the llvm should be lldb as that's the debugger (macOS)
and the llvm tools are not.

Format fixes in README.md.