Clarify gets()/fgets()/ARG_MAX in 1992/gson
By accident I stated that I changed this entry to not use gets() but
instead fgets(). This is unfortunately not true because of the way
gets() is used. I have now made it (I hope) a bit clearer. I also
pointed out that it's theoretically safe. Since it uses gets() which
comes from stdin ARG_MAX should be relevant. Thus I made the buffer size
ARG_MAX+1. But something I am not so clear on is whether or not ARG_MAX
is the entire command line or a single argument. Looking did not help
here as some say one thing and others say the other. Doing xargs
--show-limits and also checking the value of ARG_MAX suggests that it's
for the entire command but I don't know for sure of that. Nevertheless
it is safer than it was if not completely safe.