Update 2004/.gitignore
Some files that are generated were not being ignored. I note that the
configure script (from running make at the 2004/hibachi directory) does
cd build ; make but nothing is to be done and it seems to have been
compiled anyway (removing the file build/hibachi first confirms this).
But whether that's a problem or not I don't know.