Fix 2018/algmyr to not segfault on NULL file
Now it just exits 1 if a file cannot be opened. I do have a concern with
this type of fix which I will bring up on GitHub but it seems better
anyway to not crash with invalid input unless the entry should crash.
The fact the author noted that invalid paths as a limitation suggests in
this case that it is reasonable to fix. There are other limitations /
bugs the author mentions but perhaps those should not be fixed. That
remains to be seen I guess.

I can also now confirm that this entry does indeed work as expected; I
could not test it with commit 0a2c76c4a93249e352d54f4e62dd01fadc142f01
but I had to test it here to make sure the changes did not affect the
functionality.

The original source, which has a really nice layout, can be found in
prog.alt.c and is referenced in the README.md. It can be built by make
alt.