Fixes / updates to 2020/kurdyukov4
Instead of relying on curl to obtain the text files just add them to the
repo.

Typo fixes.

Formatting fixes.

Added what on Middle-earth lámatyávë is to the README.md so that people
may enjoy the award title a bit better. From Morgoth's Ring, volume X of
the wonderful set History of Middle-earth (HoMe), one of many things
that were edited and published by the late Christopher Tolkien, heir and
literary executor of the great Tolkien.