Add note about fgets() in 1990/cmills
About why it's useful. It's both a problem to keep gets() but it's also
potentially a problem to change it to fgets(). But as it will prevent
alarming the user with warnings about gets() and also causing messed up
output (see below) this is useful.

With the original code one will see under macOS something like:

    $ ./cmills.alt
    Shuffle...
    warning: this program uses gets(), which is unsafe.
    Total $1000.  Wager?

which is very wrong.

I did not delete the cmills.alt.c but perhaps it should be deleted. That
can come another time.