Add CSV files from SQL
The file auth-year-prizes.csv is in the form:

    # year_name_handle,award name

e.g.:

    2018_ferguson,Best use of weasel words
    2020_ferguson1,Don't tread on me award
    2020_ferguson2,Most enigmatic

and the person.csv file is in the form:

    # name,entry_handle,url,email,country,mastodon/twitter

e.g.:

    Cody Boone Ferguson,Cody_Boone_Ferguson,https://ioccc.xexyl.net,null,null,XX,https://fosstodon.org/@xexyl);

where 'XX' is the country (sometimes it's not just a country code but
name - that might or might not need changing).

Above I made null my email for this log entry (even though it's public
elsewhere in git). But these are just an example I'm most familiar with
(however horribly shocking that might be! :-) ) to give the idea.

I'm the only one with a Mastodon handle; the others that are not null
are twitter handles.

As far as how these were created:

    grep INSERT person.fixed.sql |sed -e 's,.*VALUES(,,g' -e "s/'//g"|awk -F, \
     '{ print $1,$2","$3","$4","","$6","$7","$8 }' |sed -e 's/NULL/null/g' -e  \
    "s/,,/,null,/g" | sort > person.csv.txt

(I changed the single sql file to several and then made some manual
fixes.)

Similar was done for the other file.

There certainly is room for improvement and it definitely needs some
fixes. This was a quick hack but it's a good start.