Restore missing 2018/README.md from 2018 preview
For some reason this was missing from the repository. I was certain that
there was one so I looked at my archive and found in the winner preview
repo for 2018 that there was indeed a README file (README.markdown).
This has been added as 2018/README.md. It felt important that this was
added back.

I did NOT update the 2018 archive tarball as I'm not sure what to even
call it - perhaps just README.markdown - but also because I'm not sure
what the status is of my archive/2018 is.

I fixed the formatting to match the style that is being used in this
repo.