Format, typo and other fixes in 1990/tbr/README.md
The typo of most importance (maybe the only typo) was that when I linked
to the archive tarball I left off /archive in the link.

Noted in the judges' comments that in the smaller version provided by
the author the exit() returns void problem is fixed but it still uses
gets(), not fgets(). I don't see the reason to do that as it'll compile
and it's really only an aside and not the entry.