Remove erroneous statement in 1990/jaw/README.md
Originally I had thought that a command in the commands to try did not
work in macOS. But I looked at the code a while back and it's only
because of the errno == 0 message in macOS and the fact the entry uses
perror(). I noted it's not a bug in bugs.md at the time but I did not
remove that part from the README.md (the part that says it doesn't
work).