Updates and fixes + add alt code in 1990/westley
Fix formatting in README.md.

Removed statement of judges that they fixed it for modern compilers that
object to 's' added to an int for short int as this is what I did and
the source that was imported had not been fixed though this might have
been a simple mistake. I had not noticed this before.

Update bugs.md: a known bug and one INABIAF (crashes on no arg).

Added alt code with the 1s as it's instructive to show what was once
allowed but is not now.