Fixes in 1991/ant/README.md
Mostly formatting fixes but one spelling fix that I feel almost guilty
about.

Though programme is correct in most cases (except US English but the
author is from Canada) with computers it's actually program. It just
seems better to have this correct if I'm going to go through and look
for typos. Besides this is a programming contest so it makes sense that
it should be correct. Easy mistake to make but should it be fixed?
Maybe, maybe not but I chose to anyway.