Reapply segfault fixes in 1998/schnitzi
After discussing it with Landon it was recommended that these fixes be
reapplied.

Updated bugs.md wrt this.

README.md has had formatting fixes as well.

Updated URL for author as it has changed. It's very little on his page
and the only page that suggests it is his is the ioccc.html page. I
couldn't find anything else.