Fix crash in 1991/westley stopping it from working
The problem was that the read only char*z[] was being written to. This
caused a crash which prevented the delightful program from working. The
fix was to change it to 'char z[][100]'. But why not a smaller number
when the longest string is < 100? Because a smaller number caused
strange output and it was easier to do than find the right value: 100 is
that big and it works well so why bother finding the smallest value? No
other reason behind it.

Fixed the shell script too so that it actually works okay (related to
'.' not being in the path in modern systems).

Updated README.md: moved the ttt.sh related things to new try section
and gave an example run (a losing example, challenging users to cause a
draw or better yet win as well as asking them to test something about
the game).

Added a fun parody to the README.md too. Check it out!

Removed this entry from bugs.md.