initial author handle and winner JSON files ideas
The `tmp/example.author_handle.json' is an initial concept for a
JSON file describing people who have won the IOCCC at least once.
If this example described a real winner, the filename would be:

```
author/author_handle.json
```

The `tmp/example.dot_winner.json` is an initial concept for a
JSON file describing an entry what won the IOCCC.
If this example described a real entry, the filename would be:

```
2024/dirname/.winner.json
```

These files are "works in progress" and are subject to change.
There is currently NO code that generates, nor manages these files.

Suggestions and questions are welcome.