Update sgit + add sgit.README.md
Add warning about use of -o n without -I depending on the sed command.

Updated version.

The README.md I'm not sure about but it feels like it might belong there
anyway. It can be removed if it does not seem useful or it is out of
context (and it might very well be). Nevertheless I think it's important
that the warning be there so it's good to update sgit here as well.