1991/westley cheating not a bug but a feature
Yes it's very easy to cheat at the game but as my late grandmother told
me when I was a kid: 'it's not cheating if you're not caught' and since
the computer does not catch you any more than her siblings did at the
card table with her hidden drawer with more cards it's not cheating! :-)
But on a more serious note it was documented as a possibility so it's
not a bug it's a feature so I added it to the README.md and bugs.md.