Add alt code to 1988/westley
The reason is it seems instructional and entertaining to see what
happens if the problem that the judges described in their remarks
occurs. Now one can do make clobber all alt; ./westley.alt; ./westley
and see if they are the same output.

The .gitignore file was updated - added westley.alt.

Updated README.md not only referring to the alt code but also format
fixes (the original fix of this commit).