Update 2011/dlowe + including colectomy on README
Remove extraneous colons.

Format fixes in README.md.

Add links to README.md.

Fix file paths in README.md.

Add try.me.sh as there are a number of commands to run in the try
section.

Add to README.md note that the URL is dead and the Internet Wayback
Machine archived it but it did not archive the scripts. This was also
added to the bugs.md file.