Rename 2020/ferguson1/cannibalism.log{,.md}
Rename cannibalism.log to cannibalism.log.md as an experiment. I
believe it'll look nicer this way. I also made some changes that should
make it easier to see where each part of the game is (in some cases for
the tail; every case for the head and insect).

Updated gameplay.md to refer to the renamed file and removed the now
irrelevant information about cannibalism.log.

I did not regenerate gameplay.html but I did not remove it either. The
reason I didn't is because there will be a tool that converts markdown
to html files and since this commit won't sync to the IOCCC website it's
not a risk in the meantime. Nevertheless I want the html file in for the
time being. The html file is still referenced in the file too because
it's preferable that it's used when everything goes live (except on
GitHub of course).