Format fixes and other updates in 2020/ferguson2
With this commit I should be able to return to the other entries though
as Landon pointed out I might never finish this and this is a reason it
might not have been as thorough as I could have been (another part is
it's long). Nevertheless the formatting should now be okay along the
lines of what was requested for all the winning entry markdown files.