Detoxify Stephen Sykes (remove drug link) :-)
Sound crazy? That's because it is! The URL is no longer owned by Stephen
and instead it's a marijuana dispensary in Canada. That's right. The
IOCCC has been endorsing drugs for some years now! :-) (or maybe it's
:-( ? :-) ).

Used Internet Wayback Machine link instead.

Updated the CSV files, the winners.html files and all the relevant
README.md files.