Address the fixed / unfixed 1996/westley situation
When I first fixed this it was because the scripts segfaulted and at
least one showed an environmental variable both of which caused some
usability issues. But I started to wonder if maybe the fixed code should
be the alt version. To do this I made a change to the scripts so one can
tell the script which version to use. I was having a problem in figuring
out how to explain it (owing to the fact I've been awake once again
since stupid o'clock and without good sleep) but before I answered that
I decided to test out the new feature to the script and the output made
me decide to at least for now (discussion can be had) keep the fix the
westley.c and the unfixed original code the alt code. This particularly
shows up in the third script where the output is hard to tell what is
part of the program and what is not. The unfixed version looks like:

    $ WESTLEY=./westley.alt ./clock3
    TERM=xterm-256color
    ./clock3: line 12: 15498 Segmentation fault: 11  ${WESTLEY} "123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\]^_\`abcdefghijkl" "/\$%&'()*+,-./\$%&'()*+,-." " ||" "
    ..3..6..9..12 Hours

     ----.----10---.----20---.----30---.----40---.----50---.---- Minutes
    "

which has a " by itself on the last line that should not be there. Then
there's the other problematic output. It should just look like:

    ..3..6..9..12 Hours

     ----.----10---.----20---.----30---.----40---.----50---.---- Minutes