Proper format fix 1987/korn/README.md
I didn't want to look at it properly (and certainly did not want to read
it) without seeing if I could figure out what it did on visual
inspection (as the judges say very few can) as there are spoilers. I was
able to so I now have no problem going through the file properly.