Locally untar all/all.tar.bz2, add all/.gitignore
This is so that it's easier to update the tarball at a later date as
it'll definitely have to be done with all the changes that we've made
(because obviously we don't want the contents of the tarball added to
the repo in this directory - that's what ../ is for - but we do want the
tarball itself in the repo as all/all.tar.bz2.

I at first had all the individual years added but that makes for the
need of manual editing the file when another year is added so instead
I've used some globs:

    1*/
    2*/
    3*/

Though of course none of us who are alive have to worry about 3*/ and
the way humans behave humans might not even get that far anyway.
Probably that's true for much of 2*/ as well! :-) (but also :-( )