Update 1996/gandalf
The URL was dead or to be more correct it required a login and password.
The Internet Wayback Machine was of no help.

Format fixes.

Removed URL from winners.html and winners.csv; updated bugs.md asking
for a current URL.

At the same time I fixed some entries in winners.csv which had too many
fields including some with a ',,'. I'm certain more are there but I
didn't want to go through the entire file now. It might be easier (since
it's temporary) to just wait for the generation of the json files and
then update those files since the CSV file is temporary anyway. The
winners.html file should probably be updated even though I believe that
will also be generated by a tool.

I'm sure there are other dead links too that need to be found and I will
do that in time.