Format and typo fix 1998/dlowe/README.md
Unfortunately the domain no longer exists as it once did (if it exists
it no longer has the script set up). Added this to bugs.md and noted
that there is a script (not checked for any security problems modernly)
if they have the bandwidth and wish to set it up.

Updated .gitignore as I added a try section.

Moved the dlowe.alt related text to new 'Alternate code' section.