Format fix 1998/dloweneil + typo fix bugs.md
The bugs.md file has been updated in this pull request because the
previous one is also about 'poot' and one of the authors of this entry
is the same. The typo has to do with the path to the C file.

I also added this entry (missing URL) to bugs.md - but referring to the
1998/dlowe entry as it's the same information.