Partially fix 2000/jarijyrki
The URL for one of the authors was no longer valid so I had to use the
Internet Wayback Machine. The winners.html and winners.csv files have
been updated for this change.

Format the authors' information.

That's all there is this time as I have to think on how to do the
formatting in a way that matches their set up (sections and numbers).
This will have to come another time.