Fixes and updates to 2000/tomx
Format fixes in README.md.

Link fix via the Internet Wayback Machine. This was also fixed in
winners.csv and winners.html.

Copy ../mkentry.c to this directory because the author made an example
of where one can modify the program to compile it instead - it felt
better than to change it to ../mkentry.c as that assumes the directory
structure is the same.

Added mkentry to .gitignore.