Update 2001/anonymous/Makefile - add cpp rule
This compiles the C pre-processed version as it seems like this will be
useful to any person crazy (by which I mean bold) enough to try and fix
it (and I give much respect to them!).

It is compiled by make all: maybe this is a mistake but it seems like
it's going to be useful if anyone is to fix it. This part can be undone
of course.