Update INABIAF and try sections in 2001/anonymous
In try section added some things to try.

In INABIAF I noted that the program will also very likely crash if the
program is not given a file name and that it will likely :-) do
something funny if used on itself (at least if compiled with 64-bit -
not sure about 32-bit).