Add alt code to 2001/dgbeards + format README.md
The alt code is based on the author's notes. It makes it faster. I
discovered a fun flaw in the change and I challenge (in the alternate
code section) others to find it!

Format fix README.md.

Update bugs.md - this program deliberately crashes if it loses (which it
is tries to do). The author provided a way to fix it. It was uuencoded
but as per discussion on GitHub spoilers are not supposed to be here any
more so I uudecoded it. This feature was also added to the INABIAF
section in the README.md.

BTW: if you're trying to lose and you lose did win or did you lose? :-)