Updates to 2001/herrmann1
Format fix README.md.

There are some missing files - added to bugs.md. Unfortunately they're
not in the archive.

I'm aware of an issue in the entry to do with more recent compilers. I
will be investigating it in the coming days very likely but it might be
said that this entry does not work at least in part or maybe in whole. I
wrote in the bugs.md that it's in part but now I think on it I think it
doesn't work at all and I'll be looking at it soon.

Updated .gitignore.

Added file that the author referred to and actually gave the contents
of. Unfortunately the missing files were not given.