Format and link fix 2004/arachnid
Fix format of README.md file. I wanted to say that the author is a
sociopath :-) for using 'w', 'a', 's' and 'd' but that's not right in
the README.md (the judges put something amusing there about it anyway)
so I'm doing it here instead. :-) Okay he's not really but the
navigation is difficult to vi users as the judges point out. Still a
very fun game and I didn't get lost!

Fix dead link via Internet Wayback Machine.

Renamed arachnid.info to arachnid.txt. I hope this helps with rendering
later on so that browsers can also view it. There are other entries
where it might have been good to consider this but I did not think of it
at the time.