Mark Ann Horton -> Mary Ann Horton
Mary Ann Horton (who gave us uuencode, contributed to BSD Unix amongst
other things) is transgender and she goes by the name Mary Ann so the
website should also say that (it had (Mark) in it which was her
birth name).

Fixed the URL as well as the old one is no longer valid.