Fix crash in 2006/toledo2 + partial format README
The problem is that 'int's were being used instead of 'FILE *'s. It did
not seem to crash in linux (at least fedora) but it did in macOS which
prevented it from working. Now it works in both. I got the feeling that
it did crash in some linux distros as well as apparently Yusuke Endoh
fixed it as well though he did it by changing the ints to longs. I'm not
sure if that would work for macOS; I fixed it prior to seeing this fix.
Some -includes were added to the Makefile too that might or might not be
relevant to the fix in macOS. It very well could be relevant though.

I did a partial format fix in README.md but I will have to finish this
tomorrow as I have to leave for the rest of the day.