Proper fix 2006/toledo2 and format README.md
I've coined a new abbreviation: IABINAF - it's a bug it's not a feature! :-)
Specifically the use of the CP/M emulator does not appear to work right.
Updated bugs.md.

As for the proper fix: it appears that not all the variables that I
changed to a FILE * can be a FILE * or else the wrong combination was
used. It turns out that using longs (like Yusuke Endoh did) will work
but in some cases FILE * is still used as technically the variables are
used as FILE *. I will very possibly look at finding the right
combination (or confirming that this is correct as is - though it seems
not to be at least not technically) another time.

Format fix README.md.