More progress with 2011/richards in macOS
I have found a problem that shows itself in both linux and macOS. In
linux it happens with both 32-bit and 64-bit but the sizes are naturally
different. It does not crash in linux however where it does in macOS.

I changed some variable types a bit as well but this still does not fix
it with macOS though they types changed are more consistent with what
they are used as.

The bugs.md has more information on what I discovered (noted above) - an
overflow.

More will have to be done on this another time in order to get it to
work with macOS.